more stable (aka non destructive reshaping) element wise operations
Former-commit-id: 78ccdcbacf1f5309938f3076bba2a17d10f16dff [formerly f36449cc4d5a54abf64cb0b92a3123b1d359f564]
Former-commit-id: 54d7ac81611dd22093149790127e032cd0a4afbf

+++ linear-algebra/linear-algebra-api/src/main/java/org/deeplearning4j/linalg/ops/transforms/Negative.java
@@ -24,7 +24,7 @@ public class Negative extends BaseElementWiseOp {
             return NDArrays.scalar(ComplexUtil.neg((org.deeplearning4j.linalg.api.complex.IComplexNumber) value.element()));
         }
         else {
-            double val = (double) value.element();
+            float val = (float) value.element();
             return NDArrays.scalar(-val);
         }
     }
